requirements.txt
deta
python-dotenv
pyyaml
lxml